pyproject.toml
[project]
name = "scrapers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crawl4ai>=0.8.0",
    "fastapi>=0.128.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
